feat: split method matching from outcome
I've split the info needed to match a method out into its own type, and then created two new interfaces: `MethodMatcherCreator`; and `ExpectationCreator`.

What this does is allows us to ensure that `Setup()` must be called with some kind of outcome (i.e. return, panic or observe), and that `Called()` can only be passed a method matching expression, not a full setup expression.