feat: add ability to match certain number of times
Added the ability to only match an expectation a certain number of times. This isn't quite finished because I still need to add full support for verifying method calls X times, but it looks like that might require a bit of refactoring first to separate the concept of method setup vs method call verification.