feat: support package local types
Added support for types that are defined in the same package as the mock. I've just gone with a simplistic approach for now of assuming that the package name won't conflict with any other imports, so it's possible an invalid mock could be generated in the situation where the mock references a type from another package with the same name.