feat: add support for a configuration file
Added support for a kelpie.yaml file to specify the mocks to generate. The advantage of this is that we can avoid re-parsing the same package multiple times if we're generating more than one mock from it.

The existing `go:generate` method of generation is still supported - this will be up to personal preference.

I've also added some extra output to the generate command to make it easier to tell if the mocks have been generated correctly or not.